new branch: refactor_dir js file src dir로 이동
