+++ src/db/mongo_db.js
@@ -1,5 +1,5 @@
 const { MongoClient } = require("mongodb");
-const config = require("../config");
+const config = require("../../config");
 
 const mongoClient = new MongoClient(config.mongo_config.uri);
