+++ src/db/mysql_db.js
@@ -1,5 +1,5 @@
 const mysql = require("mysql2/promise");
-const config = require("../config");
+const config = require("../../config");
 
 const pool = mysql.createPool(config.mysql_config);
