+++ src/service/emailService.js
@@ -1,5 +1,5 @@
 const nodemailer = require("nodemailer");
-const config = require("../config");
+const config = require("../../config");
 
 function sendEmail(to, subject, text, html) {
 	const transporter = nodemailer.createTransport(config.smtp_config);
